+++ plugins/source/azure/resources/services/costmanagement/views_fetch.go
@@ -1,0 +1,26 @@
+package costmanagement
+
+import (
+	"context"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/costmanagement/armcostmanagement"
+	"github.com/cloudquery/cloudquery/plugins/source/azure/client"
+	"github.com/cloudquery/plugin-sdk/schema"
+)
+
+func fetchViews(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- any) error {
+	cl := meta.(*client.Client)
+	svc, err := armcostmanagement.NewViewsClient(cl.Creds, cl.Options)
+	if err != nil {
+		return err
+	}
+	pager := svc.NewListByScopePager("subscriptions/"+cl.SubscriptionId, nil)
+	for pager.More() {
+		p, err := pager.NextPage(ctx)
+		if err != nil {
+			return err
+		}
+		res <- p.Value
+	}
+	return nil
+}